Make aspect_h and aspect_w doubles in EDLSession
There is no point to use float single variables on
modern computers. They are often converted silently
to doubles and back and they loose quickly accuracy
during uncareful computation sequences.
Change the types step by step as needed in the future.

+++ cinelerra/mwindow.C
@@ -2059,7 +2059,7 @@ void MWindow::save_backup()
 }
 
 
-int MWindow::create_aspect_ratio(float &w, float &h, int width, int height)
+int MWindow::create_aspect_ratio(double &w, double &h, int width, int height)
 {
 	int denominator;
 	if(!width || !height) return 1;
