+++ cinelerra/mwindow.h
@@ -284,7 +284,7 @@ public:
 	void cut();
 
 // Calculate aspect ratio from pixel counts
-	static int create_aspect_ratio(float &w, float &h, int width, int height);
+	static int create_aspect_ratio(double &w, double &h, int width, int height);
 // Calculate defaults path
 	static void create_defaults_path(char *string);
 // Compose standard window title
